parser: initial work on AST
This AST has only a few nodes:

* `Literal`

* `FloatLiteral`

* `IntegerLiteral`

* `StringLiteral`

* `CharLiteral`

* `BooleanLiteral`

* `Identifier`

Other grammar rules won't build an AST node at this time.

On `parser/src/reflect.rs` I do some weird code to turn a rule name (a
string) into a function that runs the parser, and a function that builds
the AST. I hopefully want to fix it and make it even weirder.

(yes I had to change the literal names because it's awkward to have an
AST node named `String`)